refactor: improve component name of application management pages
