+++ src/lib/components/application-management/ApplicationEditPage.svelte
