+++ src/lib/components/application-management/ApplicationListPage.svelte
